added the search bar in our components folder
