+++ package.json
@@ -12,6 +12,7 @@
     "react": "16.8.3",
     "react-dom": "16.8.3",
     "react-native": "https://github.com/expo/react-native/archive/sdk-35.0.0.tar.gz",
+    "react-native-elements": "^1.2.7",
     "react-native-web": "^0.11.7"
   },
   "devDependencies": {
